Desktop/upay_updated/upaydash/requirements.txt
# ==========================================
# UPAY Document Processing System
# Minimal Requirements (No Compilation Issues)
# Python 3.9+
# ==========================================
google-cloud-documentai

# Core Django
Django==5.2.7

# Database
psycopg2-binary==2.9.9

# REST Framework
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# API Documentation
drf-yasg==1.21.7

# CORS Headers
django-cors-headers==4.3.1

# Image Processing (pre-built wheels)
Pillow>=10.0.0

# OCR (install Tesseract separately)
pytesseract==0.3.10

# PDF to Image conversion
pdf2image==1.17.0

# Excel Generation (pure Python, no compilation)
openpyxl==3.1.2
xlsxwriter==3.1.9

# Date & Time
python-dateutil==2.8.2
pytz==2024.1

# JSON Web Token
PyJWT==2.8.0

# HTTP Requests
requests==2.31.0

# YAML Support
PyYAML==6.0.1

# Environment Variables
python-decouple==3.8

# Production Server
gunicorn==21.2.0

# Static Files
whitenoise==6.6.0

# Testing (optional)
pytest==7.4.3
pytest-django==4.7.0

# Development Tools (optional)
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Fake Data Generation (optional)
Faker==22.0.0

psycopg2-binary==2.9.9

openpyxl
drf-yasg
django-cors-headers
django-cors-headers==4.3.1 
djangorestframework
djangorestframework-simplejwt
Faker 

dj-database-url

#rm -rf .venv
# python3.11 -m venv .venv
#source .venv/bin/activate
# # Create DB tables
# python manage.py migrate

# # Create an admin user to use /admin and JWT testing
# python manage.py createsuperuser

# # (Optional) populate sample data
# python manage.py shell < api/create_sample_data.py
# python manage.py shell < api/create_sample_data_extended.py

# # Start development server
# python manage.py runserver


#loginto DB: psql -U snehbae -d upay_db
# to see all roles: \du

# npm run dev



#python migrate.py runserver